+++ app/packs/src/decidim/decidim_awesome/awesome_map/controllers/proposals_controller.js
@@ -16,7 +16,7 @@ export default class ProposalsController extends Controller {
     // add control layer for amendments if any
     if (this.awesomeMap.config.menu.amendments && this.component.amendments && !this.awesomeMap.layers.amendments) {
       this.awesomeMap.layers.amendments = {
-        label: `<span class="awesome_map-component" id="awesome_map-amendments_${this.component.id}" title="0" data-layer="amendments">${window.DecidimAwesome.texts.amendments}</span>`,
+        label: `<span class="awesome_map-component" id="awesome_map-amendments_${this.component.id}" title="0" data-layer="amendments">${window.DecidimAwesome.i18n.amendments}</span>`,
         group: new L.FeatureGroup.SubGroup(this.awesomeMap.cluster)
       }
       this.awesomeMap.controls.main.addOverlay(this.awesomeMap.layers.amendments.group, this.awesomeMap.layers.amendments.label);