+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 CHANGELOG
 =========
 
+v0.11.5
+-------
+
+Compatibility:
+  - Decidim 0.28.x
+
+Feature:
+  - Added formBuilder languages controller to avoid external CDN
+  - Fix cache hash in the global menu override
+  - Fix crash on updating empty boxes for scoped admins
+
 v0.11.4
 -------
 Compatibility: